requirement.txt
django
djangorestframework
django-filter
requests
beautifulsoup4
PyMySQL
jieba
keras
pymongo
progressbar2
h5py
bs4
lxml
django-cors-headers
mysqlclient
drfdocs


